moved quote and impersonate to heartbeats
