+++ bot/commands/all_commands.go
@@ -120,17 +120,6 @@ func (cm *CommandManager) LoadCommands() {
 			Description: "What's the latest iOS unicode bug???",
 			Cooldown:    time.Minute,
 		},
-		"/quote": {
-			Function:    cm.commandQuote,
-			Source:      CommandSourcePRIMARY,
-			Description: "Get a random quote from one of the channels.",
-			Cooldown:    time.Minute * 2,
-		},
-		"/impersonate": {
-			Function:    cm.commandImpersonate,
-			Description: "Impersonate a user using markov chains.",
-			Cooldown:    time.Minute * 10,
-		},
 	}
 	for k, v := range commands {
 		v.Settings.Cooldown = cm.Config.DefaultCooldown
