--- bot/commands/cmd_quote.go
@@ -1,87 +1,0 @@
-package commands
-
-import (
-	"fmt"
-	"strings"
-	"time"
-
-	"github.com/bwmarrin/discordgo"
-
-	"github.com/Southclaws/cj/storage"
-	"github.com/Southclaws/cj/types"
-)
-
-func (cm *CommandManager) commandQuote(
-	args string,
-	message discordgo.Message,
-	contextual bool,
-	settings types.CommandSettings,
-) (
-	context bool,
-	err error,
-) {
-	var (
-		ok          bool
-		randmessage storage.ChatLog
-		messagerr   error
-		restricted  bool
-		nick        string
-	)
-
-	restrictedChannels := [11]string{
-		"457943077789892649",
-		"536617253286707210",
-		"376371371795546112",
-		"361640153489473540",
-		"531855907374759936",
-		"531860517816500235",
-		"572476607772491786",
-		"282581181034135556",
-		"282581078643048448",
-		"415801201960026122",
-		"548332120842698753",
-	}
-
-	for ok == false {
-		restricted = false
-		randmessage, messagerr = cm.Storage.GetRandomMessage()
-		if messagerr != nil {
-			cm.Discord.ChannelMessageSend(message.ChannelID, "Could not get a quote.")
-			return
-		}
-
-		// Misses stuff like verify and doesn't allow commands to be sent either.
-		if len(randmessage.Message) > 6 && strings.Index(randmessage.Message, "/") != 0 {
-			// TODO: Do through the query.
-			for _, channel := range restrictedChannels {
-				if randmessage.DiscordChannel == channel {
-					restricted = true
-				}
-			}
-
-			if !restricted {
-				ok = true
-			}
-		}
-	}
-
-	user, membererr := cm.Discord.S.GuildMember(message.GuildID, randmessage.DiscordUserID)
-	if membererr != nil || len(user.Nick) == 0 {
-		// No longer on the server or has no nickname.
-		user, membererr := cm.Discord.S.User(randmessage.DiscordUserID)
-		if membererr != nil {
-			cm.Discord.ChannelMessageSend(message.ChannelID, "Failed to get user.")
-			err = membererr
-			return
-		}
-
-		nick = user.Username
-	} else {
-		nick = user.Nick
-	}
-
-	time := time.Unix(randmessage.Timestamp, 0)
-	cm.Discord.ChannelMessageSend(message.ChannelID, fmt.Sprintf("\"%s\" ~ **%s**, **%s %d**", randmessage.Message, nick,
-		time.Month().String(), time.Year()))
-	return
-}
